Split files into lib/crabstone/
