+++ crabstone.gemspec
@@ -1,27 +1,32 @@
-require 'rubygems'
+lib = File.expand_path('../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+
+require 'date'
+
+require 'crabstone/version'
 
 Gem::Specification.new do |spec|
   spec.name       = 'crabstone'
-  spec.version    = '3.0.3'
-  spec.author     = 'Ben Nagy'
-  spec.license    = 'BSD'
-  spec.email      = 'crabstone@ben.iagu.net'
-  spec.homepage   = 'https://github.com/bnagy/crabstone'
+  spec.version    = Crabstone::VERSION
+  spec.date       = Date.today.to_s
+  spec.authors    = ['Ben Nagy', 'david942j']
+  spec.license    = 'BSD-3-Clause'
+  spec.email      = ['crabstone@ben.iagu.net', 'david942j@gmail.com']
+  spec.homepage   = 'https://github.com/david942j/crabstone'
   spec.summary    = 'Ruby FFI bindings for the capstone disassembly engine'
-  spec.test_files = Dir['test/*.rb']
-  spec.files      = Dir['**/*'].delete_if{ |item| item.include?('git') }
+  spec.files      = Dir['lib/**/*.rb']
 
   spec.extra_rdoc_files = ['CHANGES.md', 'README.md', 'MANIFEST']
 
   spec.add_runtime_dependency 'ffi' unless RUBY_PLATFORM =~/java/
   spec.add_development_dependency 'test-unit'
 
-  spec.description = <<-EOF
+  spec.description = <<-EOS
 
   Capstone is a disassembly engine written by Nguyen Anh Quynh, available here
   https://github.com/aquynh/capstone. This is the Ruby FFI binding. We test
   against MRI 2.0.0,  2.1.0 and JRuby 1.7.8. AFAIK it works with rubinius
   2.2.1.
 
-    EOF
+  EOS
 end
