+++ lib/crabstone.rb
@@ -3,28 +3,24 @@
 # Additional binding work by Ben Nagy
 # (c) 2013 COSEINC. All Rights Reserved.
 
-require 'ffi'
-
-require_relative 'arch/x86'
-require_relative 'arch/x86_registers'
-require_relative 'arch/arm'
-require_relative 'arch/arm_registers'
-require_relative 'arch/arm64'
-require_relative 'arch/arm64_registers'
-require_relative 'arch/mips'
-require_relative 'arch/mips_registers'
-require_relative 'arch/ppc'
-require_relative 'arch/ppc_registers'
-require_relative 'arch/sparc'
-require_relative 'arch/sparc_registers'
-require_relative 'arch/systemz'
-require_relative 'arch/sysz_registers'
-require_relative 'arch/xcore'
-require_relative 'arch/xcore_registers'
+require_relative 'crabstone/arch/x86'
+require_relative 'crabstone/arch/x86_registers'
+require_relative 'crabstone/arch/arm'
+require_relative 'crabstone/arch/arm_registers'
+require_relative 'crabstone/arch/arm64'
+require_relative 'crabstone/arch/arm64_registers'
+require_relative 'crabstone/arch/mips'
+require_relative 'crabstone/arch/mips_registers'
+require_relative 'crabstone/arch/ppc'
+require_relative 'crabstone/arch/ppc_registers'
+require_relative 'crabstone/arch/sparc'
+require_relative 'crabstone/arch/sparc_registers'
+require_relative 'crabstone/arch/systemz'
+require_relative 'crabstone/arch/sysz_registers'
+require_relative 'crabstone/arch/xcore'
+require_relative 'crabstone/arch/xcore_registers'
 
 module Crabstone
-
-  VERSION = '3.0.3'
 
   # API version
   BINDING_MAJ = 3
@@ -113,452 +109,9 @@ module Crabstone
     true  => 3, #trololol
     false => 0
   }
-
-  class ErrArch < StandardError; end
-  class ErrCsh < StandardError; end
-  class ErrHandle < StandardError; end
-  class ErrMem < StandardError; end
-  class ErrMode < StandardError; end
-  class ErrOK < StandardError; end
-  class ErrOption < StandardError; end
-  class ErrDetail < StandardError; end
-  class ErrMemSetup < StandardError; end
-  class ErrVersion < StandardError; end
-  class ErrDiet < StandardError; end
-  class ErrSkipData < StandardError; end
-  class ErrX86ATT < StandardError; end
-  class ErrX86Intel < StandardError; end
-
-  ERRNO = {
-    0  => ErrOK,
-    1  => ErrMem,
-    2  => ErrArch,
-    3  => ErrHandle,
-    4  => ErrCsh,
-    5  => ErrMode,
-    6  => ErrOption,
-    7  => ErrDetail,
-    8  => ErrMemSetup,
-    9  => ErrVersion,
-    10 => ErrDiet,
-    11 => ErrSkipData,
-    12 => ErrX86ATT,
-    13 => ErrX86Intel,
-  }
-
-  ERRNO_KLASS = ERRNO.invert
-
-  def self.raise_errno errno
-    err_klass = ERRNO[errno]
-    raise RuntimeError, "Internal Error: Tried to raise unknown errno" unless err_klass
-    err_str = Binding.cs_strerror(errno)
-    raise err_klass, err_str
-  end
-
-  module Binding
-
-    extend FFI::Library
-    ffi_lib ['capstone', 'libcapstone.so.3']
-
-    # This is because JRuby FFI on x64 Windows thinks size_t is 32 bit
-    case FFI::Platform::ADDRESS_SIZE
-    when 64
-      typedef :ulong_long, :size_t
-    when 32
-      typedef :ulong, :size_t
-    else
-      fail "Unsupported native address size"
-    end
-
-    typedef :size_t, :csh
-    typedef :size_t, :cs_opt_value
-    typedef :uint, :cs_opt_type
-    typedef :uint, :cs_err
-    typedef :uint, :cs_arch
-    typedef :uint, :cs_mode
-
-    class Architecture < FFI::Union
-      layout(
-        :arm, ARM::Instruction,
-        :arm64, ARM64::Instruction,
-        :mips, MIPS::Instruction,
-        :x86, X86::Instruction,
-        :ppc, PPC::Instruction,
-        :sparc, Sparc::Instruction,
-        :sysz, SysZ::Instruction,
-        :xcore, XCore::Instruction
-      )
-    end
-
-    class Detail < FFI::Struct
-      layout(
-        :regs_read, [:uint8, 12],
-        :regs_read_count, :uint8,
-        :regs_write, [:uint8, 20],
-        :regs_write_count, :uint8,
-        :groups, [:uint8, 8],
-        :groups_count, :uint8,
-        :arch, Architecture
-      )
-    end
-
-    class Instruction < FFI::ManagedStruct
-      layout(
-        :id, :uint,
-        :address, :ulong_long,
-        :size, :uint16,
-        :bytes, [:uchar, 16],
-        :mnemonic, [:char, 32],
-        :op_str, [:char, 160],
-        :detail, Detail.by_ref
-      )
-
-      def self.release(ptr)
-        detail_ptr = ptr.+(Instruction.offset_of(:detail)).read_ptr
-        Binding.free(detail_ptr)
-        Binding.free(ptr)
-      end
-    end
-
-    callback :skipdata_cb, [:pointer, :size_t, :size_t, :pointer], :size_t
-
-    class SkipdataConfig < FFI::Struct
-      layout(
-        :mnemonic, :pointer,
-        :callback, :skipdata_cb,
-        :unused, :pointer
-      )
-    end
-
-    attach_function(
-      :cs_disasm,
-      [:csh, :pointer, :size_t, :ulong_long, :size_t, :pointer],
-      :size_t
-    )
-    attach_function :cs_close, [:pointer], :cs_err
-    attach_function :cs_errno, [:csh], :cs_err
-    attach_function :cs_group_name, [:csh, :uint], :string
-    attach_function :cs_insn_group, [:csh, Instruction, :uint], :bool
-    attach_function :cs_insn_name, [:csh, :uint], :string
-    attach_function :cs_op_count, [:csh, Instruction, :uint], :cs_err
-    attach_function :cs_open, [:cs_arch, :cs_mode, :pointer], :cs_err
-    attach_function :cs_option, [:csh, :cs_opt_type, :cs_opt_value], :cs_err
-    attach_function :cs_reg_name, [:csh, :uint], :string
-    attach_function :cs_reg_read, [:csh, Instruction, :uint], :bool
-    attach_function :cs_reg_write, [:csh, Instruction, :uint], :bool
-    attach_function :cs_strerror, [:cs_err], :string
-    attach_function :cs_support, [:cs_arch], :bool
-    attach_function :cs_version, [:pointer, :pointer], :uint
-    attach_function :memcpy, [:pointer, :pointer, :size_t], :pointer
-    attach_function :malloc, [:size_t], :pointer
-    attach_function :free, [:pointer], :void
-
-  end # Binding
-
-  # This is a C engine build option, so we can set it here, not when we
-  # instantiate a new Disassembler.
-  DIET_MODE = Binding.cs_support SUPPORT_DIET
-  # Diet mode means:
-  # - No op_str or mnemonic in Instruction
-  # - No regs_read, regs_write or groups ( even with detail on )
-  # - No reg_name or insn_name id2str convenience functions
-  # - detail mode CAN still be on - so the arch insn operands MAY be available
-
-  class Instruction
-
-    attr_reader :arch, :csh, :raw_insn
-
-    ARCHS = {
-      arm: ARCH_ARM,
-      arm64: ARCH_ARM64,
-      x86: ARCH_X86,
-      mips: ARCH_MIPS,
-      ppc: ARCH_PPC,
-      sparc: ARCH_SPARC,
-      sysz: ARCH_SYSZ,
-      xcore: ARCH_XCORE
-    }.invert
-
-    ARCH_CLASSES = {
-      ARCH_ARM   => ARM,
-      ARCH_ARM64 => ARM64,
-      ARCH_X86   => X86,
-      ARCH_MIPS  => MIPS,
-      ARCH_PPC   => PPC,
-      ARCH_SPARC => Sparc,
-      ARCH_SYSZ  => SysZ,
-      ARCH_XCORE => XCore
-    }
-
-    def initialize csh, insn, arch
-      @arch       = arch
-      @csh        = csh
-      @raw_insn   = insn
-      if detailed?
-        @detail     = insn[:detail]
-        @arch_insn  = @detail[:arch][ARCHS[arch]]
-        @regs_read  = @detail[:regs_read].first( @detail[:regs_read_count] )
-        @regs_write = @detail[:regs_write].first( @detail[:regs_write_count] )
-        @groups     = @detail[:groups].first( @detail[:groups_count] )
-      end
-    end
-
-    def name
-      raise_if_diet
-      name = Binding.cs_insn_name(csh, id)
-      Crabstone.raise_errno( ERRNO_KLASS[ErrCsh] ) unless name
-      name
-    end
-
-    def group_name grp
-      raise_if_diet
-      name = Binding.cs_group_name(csh, Integer(grp))
-      Crabstone.raise_errno( ERRNO_KLASS[ErrCsh] ) unless name
-      name
-    end
-
-    # It's more informative to raise if CS_DETAIL is off than just return nil
-    def detailed?
-      not @raw_insn[:detail].pointer.null?
-    end
-
-    def detail
-      raise_unless_detailed
-      @detail
-    end
-
-    def regs_read
-      raise_unless_detailed
-      raise_if_diet
-      @regs_read
-    end
-
-    def regs_write
-      raise_unless_detailed
-      raise_if_diet
-      @regs_write
-    end
-
-    def groups
-      raise_unless_detailed
-      raise_if_diet
-      @groups
-    end
-
-    def group? groupid
-      raise_unless_detailed
-      raise_if_diet
-      Binding.cs_insn_group csh, raw_insn, groupid
-    end
-
-    def reads_reg? reg
-      raise_unless_detailed
-      raise_if_diet
-      Binding.cs_reg_read csh, raw_insn, ARCH_CLASSES[arch].register( reg )
-    end
-
-    def writes_reg? reg
-      raise_unless_detailed
-      raise_if_diet
-      Binding.cs_reg_write csh, raw_insn, ARCH_CLASSES[arch].register( reg )
-    end
-
-    def mnemonic
-      raise_if_diet
-      raw_insn[:mnemonic]
-    end
-
-    def op_str
-      raise_if_diet
-      raw_insn[:op_str]
-    end
-
-    def op_count op_type=nil
-      raise_unless_detailed
-      if op_type
-        Binding.cs_op_count csh, raw_insn, op_type
-      else
-        self.operands.size
-      end
-    end
-
-    def bytes
-      raw_insn[:bytes].first raw_insn[:size]
-    end
-
-    # So an Instruction should respond to all the methods in Instruction, and
-    # all the methods in the Arch specific Instruction class. The methods /
-    # members that have special handling for detail mode or diet mode are
-    # handled above. The rest is dynamically dispatched below.
-    def method_missing meth, *args
-      if raw_insn.members.include? meth
-        # Dispatch to toplevel Instruction class ( this file )
-        raw_insn[meth]
-      else
-        # Nothing else is available without details.
-        if not detailed?
-          raise(
-            NoMethodError,
-            "Either CS_DETAIL is off, or #{self.class} doesn't implement #{meth}"
-          )
-        end
-        # Dispatch to the architecture specific Instruction ( in arch/ )
-        if @arch_insn.respond_to? meth
-          @arch_insn.send meth, *args
-        elsif @arch_insn.members.include? meth
-          @arch_insn[meth]
-        else
-          raise NoMethodError, "Unknown method #{meth} for #{self.class}"
-        end
-      end
-    end
-
-    private
-
-    def raise_unless_detailed
-      Crabstone.raise_errno( Crabstone::ERRNO_KLASS[ErrDetail] ) unless detailed?
-    end
-
-    def raise_if_diet
-      Crabstone.raise_errno( Crabstone::ERRNO_KLASS[ErrDiet] ) if DIET_MODE
-    end
-
-  end
-
-
-  class Disassembler
-
-    attr_reader :arch, :mode, :csh, :syntax, :decomposer
-
-    def initialize arch, mode
-
-      maj, min = version
-      if maj != BINDING_MAJ || min != BINDING_MIN
-        raise "FATAL: Binding for #{BINDING_MAJ}.#{BINDING_MIN}, found #{maj}.#{min}"
-      end
-
-      @arch    = arch
-      @mode    = mode
-      p_size_t = FFI::MemoryPointer.new :ulong_long
-      @p_csh    = FFI::MemoryPointer.new p_size_t
-      if ( res = Binding.cs_open( arch, mode, @p_csh )).nonzero?
-        Crabstone.raise_errno res
-      end
-
-      @csh = @p_csh.read_ulong_long
-
-    end
-
-    # After you close the engine, don't use it anymore. Can't believe I even
-    # have to write this.
-    def close
-      if ( res = Binding.cs_close(@p_csh) ).nonzero?
-        Crabstone.raise_errno res
-      end
-    end
-
-    def syntax= new_stx
-      Crabstone.raise_errno( Crabstone::ERRNO_KLASS[ErrOption] ) unless SYNTAX[new_stx]
-      res = Binding.cs_option(csh, OPT_SYNTAX, SYNTAX[new_stx])
-      Crabstone.raise_errno res if res.nonzero?
-      @syntax = new_stx
-    end
-
-    def decomposer= new_val
-      res = Binding.cs_option(csh, OPT_DETAIL, DETAIL[!!(new_val)])
-      Crabstone.raise_errno res if res.nonzero?
-      @decomposer = !!(new_val)
-    end
-
-    def version
-      maj = FFI::MemoryPointer.new(:int)
-      min = FFI::MemoryPointer.new(:int)
-      Binding.cs_version maj, min
-      [ maj.read_int, min.read_int ]
-    end
-
-    def diet?
-      DIET_MODE
-    end
-
-    def errno
-      Binding.cs_errno(csh)
-    end
-
-    def skipdata mnemonic='.byte'
-
-      cfg = Binding::SkipdataConfig.new
-      cfg[:mnemonic] = FFI::MemoryPointer.from_string String(mnemonic)
-
-      if block_given?
-
-        real_cb = FFI::Function.new(
-          :size_t,
-          [:pointer, :size_t, :size_t, :pointer]
-        ) {|code, sz, offset, _|
-
-          code = code.read_array_of_uchar(sz).pack('c*')
-          begin
-            res = yield code, offset
-            Integer(res)
-          rescue
-            warn "Error in skipdata callback: #{$!}"
-            # It will go on to crash, but now at least there's more info :)
-          end
-        }
-
-        cfg[:callback] = real_cb
-
-      end
-
-      res = Binding.cs_option(csh, OPT_SKIPDATA_SETUP, cfg.pointer.address)
-      Crabstone.raise_errno res if res.nonzero?
-      res = Binding.cs_option(csh, OPT_SKIPDATA, SKIPDATA[true])
-      Crabstone.raise_errno res if res.nonzero?
-    end
-
-    def skipdata_off
-      res = Binding.cs_option(csh, OPT_SKIPDATA, SKIPDATA[false])
-      Crabstone.raise_errno res if res.nonzero?
-    end
-
-    def reg_name regid
-      Crabstone.raise_errno( Crabstone::ERRNO_KLASS[ErrDiet] ) if DIET_MODE
-      name = Binding.cs_reg_name(csh, regid)
-      Crabstone.raise_errno( Crabstone::ERRNO_KLASS[ErrCsh] ) unless name
-      name
-    end
-
-    def disasm code, offset, count = 0
-      return [] if code.empty?
-
-      insn_ptr   = FFI::MemoryPointer.new :pointer
-      insn_count = Binding.cs_disasm(
-        @csh,
-        code,
-        code.bytesize,
-        offset,
-        count,
-        insn_ptr
-      )
-      Crabstone.raise_errno(errno) if insn_count.zero?
-
-      insns = (0...insn_count * Binding::Instruction.size).step(Binding::Instruction.size).map do |off|
-        cs_insn_ptr = Binding.malloc Binding::Instruction.size
-        cs_insn = Binding::Instruction.new cs_insn_ptr
-        Binding.memcpy(cs_insn_ptr, insn_ptr.read_pointer + off, Binding::Instruction.size)
-        Instruction.new @csh, cs_insn, @arch
-      end
-
-      Binding.free(insn_ptr.read_pointer)
-
-      insns
-    end
-
-    def set_raw_option opt, val
-      res = Binding.cs_option csh, opt, val
-      Crabstone.raise_errno res if res.nonzero?
-    end
-
-  end
 end
+
+require 'crabstone/binding'
+require 'crabstone/disassembler'
+require 'crabstone/error'
+require 'crabstone/version'
