+++ lib/crabstone/arch/arm.rb
