+++ lib/crabstone/arch/arm_const.rb
