+++ lib/crabstone/arch/arm_registers.rb
