+++ lib/crabstone/arch/mips.rb
