+++ lib/crabstone/arch/mips_const.rb
