+++ lib/crabstone/arch/mips_registers.rb
