+++ lib/crabstone/arch/ppc.rb
