+++ lib/crabstone/arch/ppc_const.rb
