+++ lib/crabstone/arch/ppc_registers.rb
