+++ lib/crabstone/arch/sparc_const.rb
