+++ lib/crabstone/arch/sparc_registers.rb
