+++ lib/crabstone/arch/x86.rb
