+++ lib/crabstone/arch/x86_const.rb
