+++ lib/crabstone/arch/x86_registers.rb
