+++ lib/crabstone/arch/xcore.rb
