+++ lib/crabstone/arch/xcore_const.rb
