+++ lib/crabstone/arch/xcore_registers.rb
