+++ lib/crabstone/version.rb
@@ -1,0 +1,3 @@
+module Crabstone
+  VERSION = '3.0.3'.freeze
+end
